Restructured code, improved readability, deleted junk
