--- CHANGELOG
@@ -1,38 +1,0 @@
-v1.8.0
-	*
-
-v1.7.3
-	* Renamed node-htmlparser.* to htmlparser.* and created shims for people still expecting node-htmlparser.*
-
-v1.7.2
-	* Document position feature fixed to work correctly with chunked parsing
-
-v1.7.1
-	* Document position feature disabled until it works correctly with chunked parsing
-
-v1.7.0
-	* Empty tag checking switch to being case insensitive [fgnass]
-	* Added feature to include document position (row, col) in element data [fgnass]
-	* Added parser option "includeLocation" to enable document position data
-
-v1.6.4
-	* Fixed 'prevElement' error [Swizec]
-
-v1.6.3
-	* Updated to support being an npm package
-	* Fixed DomUtils.testElement()
-
-v1.6.1
-	* Optimized DomUtils by up to 2-3x 
-
-v1.6.0
-	* Added support for RSS/Atom feeds
-
-v1.5.0
-	* Added DefaultHandler option "enforceEmptyTags" so that XML can be parsed correctly
-
-v1.4.2
-	* Added tests for parsing XML with namespaces
-
-v1.4.1
-	* Added minified version
